runc: Allow values "private" and "slave" for rootfsPropagation
Allow rootfsPropagation to be either "private" or "slave" and mount rootfs
accordingly. Default continues to be "private".

This will allow user to opt-in container rootfs mount propagation mode to
be RSLAVE (recursive slave), and that means that any mounts done in parent
namespace can propagate in container mount namespace.

This patch is inspired from Huamin Chen's patch.

Signed-off-by: Vivek Goyal <vgoyal@redhat.com>